requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
attrs==23.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.12.2
cachelib==0.10.2
certifi @ file:///C:/ci/certifi_1655889035820/work/certifi
charset-normalizer==3.2.0
click==8.1.7
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dataclasses-json==0.5.14
debugpy @ file:///C:/ci/debugpy_1637091942146/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
diskcache==5.6.3
docker==4.2.0
entrypoints @ file:///C:/ci/entrypoints_1649908465568/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
FLAML==2.1.1
Flask==2.2.2
Flask-Session==0.5.0
fonttools==4.25.0
freetype-py==2.4.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
ipykernel @ file:///C:/ci/ipykernel_1647000950647/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1651602038898/work
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644297245519/work
Jinja2==3.1.2
joblib==1.3.2
jupyter-client @ file:///C:/ci/jupyter_client_1650534160968/work
jupyter-core @ file:///C:/ci/jupyter_core_1651656286245/work
kiwisolver @ file:///C:/ci/kiwisolver_1653274190210/work
langchain==0.0.292
langsmith==0.0.37
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib @ file:///C:/ci/matplotlib-suite_1647441845039/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186172375/work
mkl-service==2.4.0
multidict==6.0.4
munkres==1.1.4
mypy-extensions==1.0.0
nest-asyncio @ file:///C:/ci/nest-asyncio_1649848092662/work
numexpr==2.8.6
numpy @ file:///C:/ci/numpy_and_numpy_base_1653574883459/work
openai==0.28.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.5.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.0.0
playwright==1.37.0
prettytable==3.9.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyautogen==0.1.10
pydantic==2.3.0
pydantic_core==2.6.3
pyecharts==2.0.4
pyee==9.0.4
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pypiwin32==223
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq @ file:///C:/ci/pyzmq_1638417001829/work
requests==2.31.0
scikit-learn==1.3.1
scipy==1.9.1
simplejson==3.19.1
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post9
soupsieve==2.5
SQLAlchemy==2.0.20
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.2.0
tornado @ file:///C:/ci/tornado_1606942375796/work
tqdm==4.66.1
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websocket-client==1.6.4
Werkzeug==2.3.7
wincertstore==0.2
yarl==1.9.2
zipp==3.16.2
